Better handling of 'unavailable' property
As per the official docs (https://developers.home-assistant.io/docs/en/integration_quality_scale_index.html#silver), "Handles device/service unavailable. Log a warning once when unavailable, log once when reconnected."

So I modified the connect funcion to only warn about an error if it has not been logged before.
Also, even if the device is unavailable, the logs would be filled with 'Update taking longer than 10s' every updates, to prevent that I added a timeout to ADB's connect function, and now everything is working as expected and the logs are not flooded. Also without this timeout, it would cause a delay of ~20s while launching HA.