Improve the device_info function
As a user on the forum indicated, the output of adb shell getprop is not the same on different Android TV devices. Indeed on the Shield there is no wifi or bluetooth mac address.
I implemented a new way to fetch the Wifi mac if it's not present in the getprop output, and also removed the bluetooth mac as it's actually not needed.